rename courtDecision file using camelCase
