+++ backend/src/modules/courtDecision/courtDecision.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CourtDecisionResolver } from './court-decision.resolver';
-import { CourtDecisionService } from './court-decision.service';
+import { CourtDecisionResolver } from './courtDecision.resolver';
+import { CourtDecisionService } from './courtDecision.service';
 
 @Module({
   providers: [CourtDecisionResolver, CourtDecisionService]
